Revert "done"
This reverts commit 817f262b7b41e62ea36e58b02f00dccfdf14fba8.

+++ exercises/23_conversions/using_as.rs
@@ -10,11 +10,11 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM  DONE
+// I AM NOT DONE
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len() as f64
+    total / values.len()
 }
 
 fn main() {
